Refactor: clean up top directory
