--- src/app/(after-login)/(top)/exam-detail/[examId]/exam-detail-api.ts
@@ -1,12 +1,0 @@
-export async function readExamDetail(examId: string) {
-  try {
-    const response = await fetch(`/api/exam/${examId}`, {
-      method: 'GET',
-    });
-
-    const json = await response.json();
-    return json.data;
-  } catch (e) {
-    console.error(`[createQuestion] failed by ${e}`);
-  }
-}
